+++ tests/test_examples.py
@@ -91,7 +91,7 @@ async def test_latex_example(path_tmp: Path, name: str):
     await run_example(Path("tests/examples") / name, path_tmp, OVERWRITE_EXPECTED)
 
 
-def has_tectonic_0_16_x():
+def has_tectonic_0_17_x():
     if not shutil.which("tectonic"):
         return False
     cp = subprocess.run(
@@ -102,10 +102,10 @@ def has_tectonic_0_16_x():
         check=True,
         text=True,
     )
-    return cp.stdout.split()[-1].startswith("0.16.")
+    return cp.stdout.split()[-1].startswith("0.17.")
 
 
-@pytest.mark.skipif(not has_tectonic_0_16_x(), reason="No Tectonic 0.16.x")
+@pytest.mark.skipif(not has_tectonic_0_17_x(), reason="No Tectonic 0.17.x")
 @pytest.mark.parametrize(
     "name",
     [